added share and save button , onclick listeners not implemented
